+++ app/src/main/java/multicampus/project/multigo/ui/basket/BasketRecyclerViewAdapter.java
@@ -1,32 +1,78 @@
 package multicampus.project.multigo.ui.basket;
 
-import androidx.recyclerview.widget.RecyclerView;
-
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.Button;
 import android.widget.TextView;
 
-import multicampus.project.multigo.R;
-import multicampus.project.multigo.ui.basket.BasketFragment.OnListFragmentInteractionListener;
-import multicampus.project.multigo.data.ItemsVO;
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+import androidx.recyclerview.widget.RecyclerView;
 
+import com.google.firebase.database.ChildEventListener;
+import com.google.firebase.database.DataSnapshot;
+import com.google.firebase.database.DatabaseError;
+import com.google.firebase.database.DatabaseReference;
+
+import java.util.ArrayList;
 import java.util.List;
 
+import multicampus.project.multigo.R;
+import multicampus.project.multigo.data.BasketItemVO;
+import multicampus.project.multigo.ui.basket.BasketFragment.OnListFragmentInteractionListener;
+
 /**
- * {@link RecyclerView.Adapter} that can display a {@link ItemsVO} and makes a call to the
+ * {@link RecyclerView.Adapter} that can display a {@link BasketItemVO} and makes a call to the
  * specified {@link OnListFragmentInteractionListener}.
  * TODO: Replace the implementation with code for your data type.
  */
 public class BasketRecyclerViewAdapter extends RecyclerView.Adapter<BasketRecyclerViewAdapter.ViewHolder> {
 
-    private final List<ItemsVO> mValues;
+    private List<BasketItemVO> mBasketItems = new ArrayList<>();
+    private List<String> mItemsId = new ArrayList<>();
     private final OnListFragmentInteractionListener mListener;
+    private DatabaseReference mDatabaseRef;
+    private ChildEventListener mChildEventListener;
 
-    public BasketRecyclerViewAdapter(List<ItemsVO> items, OnListFragmentInteractionListener listener) {
-        mValues = items;
+    public BasketRecyclerViewAdapter(OnListFragmentInteractionListener listener, DatabaseReference ref) {
         mListener = listener;
+        mDatabaseRef = ref;
+
+        ChildEventListener childEventListener = new ChildEventListener() {
+            @Override
+            public void onChildAdded(@NonNull DataSnapshot dataSnapshot, @Nullable String s) {
+                BasketItemVO item = dataSnapshot.getValue(BasketItemVO.class);
+
+                mItemsId.add(dataSnapshot.getKey());
+                mBasketItems.add(item);
+                notifyItemInserted(mBasketItems.size() - 1);
+            }
+
+            @Override
+            public void onChildChanged(@NonNull DataSnapshot dataSnapshot, @Nullable String s) {
+
+            }
+
+            @Override
+            public void onChildRemoved(@NonNull DataSnapshot dataSnapshot) {
+
+            }
+
+            @Override
+            public void onChildMoved(@NonNull DataSnapshot dataSnapshot, @Nullable String s) {
+
+            }
+
+            @Override
+            public void onCancelled(@NonNull DatabaseError databaseError) {
+
+            }
+        };
+
+        ref.addChildEventListener(childEventListener);
+
+        mChildEventListener = childEventListener;
     }
 
     @Override
@@ -38,10 +84,10 @@ public class BasketRecyclerViewAdapter extends RecyclerView.Adapter<BasketRecycl
 
     @Override
     public void onBindViewHolder(final ViewHolder holder, int position) {
-        holder.mItem = mValues.get(position);
-        holder.mNameView.setText(mValues.get(position).getName());
+        holder.mItem = mBasketItems.get(position);
+        holder.mNameView.setText(mBasketItems.get(position).getName());
         holder.mNumberView.setText("1");
-        holder.mPriceView.setText(String.valueOf(mValues.get(position).getPrice()));
+        holder.mPriceView.setText(String.valueOf(mBasketItems.get(position).getPrice()));
 
         holder.mView.setOnClickListener(v -> {
             if (null != mListener) {
@@ -54,7 +100,7 @@ public class BasketRecyclerViewAdapter extends RecyclerView.Adapter<BasketRecycl
 
     @Override
     public int getItemCount() {
-        return mValues.size();
+        return mBasketItems.size();
     }
 
     public class ViewHolder extends RecyclerView.ViewHolder {
@@ -63,7 +109,7 @@ public class BasketRecyclerViewAdapter extends RecyclerView.Adapter<BasketRecycl
         final TextView mPriceView;
         final TextView mNumberView;
         final Button mDeleteBtn;
-        ItemsVO mItem;
+        BasketItemVO mItem;
 
         ViewHolder(View view) {
             super(view);
@@ -77,6 +123,12 @@ public class BasketRecyclerViewAdapter extends RecyclerView.Adapter<BasketRecycl
         @Override
         public String toString() {
             return super.toString() + " '" + mNumberView.getText() + "'";
+        }
+    }
+
+    public void cleanupListener(){
+        if(mChildEventListener != null){
+            mDatabaseRef.removeEventListener(mChildEventListener);
         }
     }
 }
